Fix _read_data function to cope with binary file uploads
raw_data is a bytes object (rather than a str object), which can contain arbitrary binary data (e.g. an mp3 file) and therefore calling .deocde('utf8') on it can just cause it to blow up, depending on what bytes happen to be in the data. I think that this is the correct way to handle the data here.